Major restructuring and discarded all output code. Parser works.
