+++ Makefile
@@ -1,13 +1,17 @@
 NAME = ft_printf
 
 SOURCES = ft_printf.c \
-	strany.c \
-	parse_format.c \
-	output.c \
-	utoa.c \
-	ft_ftoa.c \
-	binary.c \
-	output_helper.c \
+	strany.c          \
+	parse_format.c    \
+	output_string.c   \
+	output_pointer.c  \
+	output_float.c    \
+	output_uint.c     \
+	output_int.c      \
+	utoa.c            \
+	ft_ftoa.c         \
+	binary.c          \
+	test.c
 
 OBJECTS = $(subst .c,.o,$(SOURCES))
